+++ src/mobx/model/network/eth.network.ts
@@ -1,4 +1,4 @@
-import { NETWORK_IDS, NETWORK_LIST } from 'config/constants';
+import { NETWORK_IDS } from 'config/constants';
 import { createChainBatchConfig, toSettConfig } from 'web3/config/config-utils';
 import { BatchCallRequest } from 'web3/interface/batch-call-request';
 import { Deploy } from 'web3/interface/deploy';
@@ -11,15 +11,17 @@ import { SettState } from '../setts/sett-state';
 import { toRecord } from 'web3/config/token-config';
 import { ProtocolTokens } from 'web3/interface/protocol-token';
 import { FLAGS } from 'config/environment';
+import { ChainNetwork } from 'config/enums/chain-network.enum';
+import { Currency } from 'config/enums/currency.enum';
 
 export class Ethereum extends Network {
 	constructor() {
 		super(
 			'https://etherscan.io',
 			'Ethereum',
-			NETWORK_LIST.ETH,
+			ChainNetwork.Ethereum,
 			NETWORK_IDS.ETH,
-			'ETH',
+			Currency.ETH,
 			ETH_DEPLOY,
 			ethSettDefinitions,
 		);
@@ -58,11 +60,19 @@ export class Ethereum extends Network {
 
 	batchRequests(setts: SettMap, address: string): BatchCallRequest[] {
 		const tokenAddresses = Object.values(setts).map((sett) => sett.underlyingToken);
+		const nonSettTokenAddresses = [deploy.digg_system.DROPT['DROPT-3'].longToken];
 		const settAddresses = Object.values(setts).map((sett) => sett.vaultToken);
 		const generalSetts = settAddresses.filter((sett) => setts[sett].state === SettState.Open);
 		const guardedSetts = settAddresses.filter((sett) => setts[sett].state !== SettState.Open);
 		const geyserAddresses = ethSetts.map((sett) => sett.geyser).filter((geyser): geyser is string => !!geyser);
-		return createChainBatchConfig(tokenAddresses, generalSetts, guardedSetts, geyserAddresses, address);
+		return createChainBatchConfig(
+			tokenAddresses,
+			nonSettTokenAddresses,
+			generalSetts,
+			guardedSetts,
+			geyserAddresses,
+			address,
+		);
 	}
 
 	async updateGasPrices(): Promise<GasPrices> {