+++ src/mobx/model/network/matic.network.ts
@@ -1,4 +1,4 @@
-import { NETWORK_IDS, NETWORK_LIST } from 'config/constants';
+import { NETWORK_IDS } from 'config/constants';
 import { toRecord } from 'web3/config/token-config';
 import { Deploy } from 'web3/interface/deploy';
 import { ProtocolTokens } from 'web3/interface/protocol-token';
@@ -6,15 +6,17 @@ import { GasPrices } from '../system-config/gas-prices';
 import { BadgerSett } from '../vaults/badger-sett';
 import { Network } from './network';
 import deploy from '../../../config/deployments/matic.json';
+import { ChainNetwork } from 'config/enums/chain-network.enum';
+import { Currency } from 'config/enums/currency.enum';
 
 export class Polygon extends Network {
 	constructor() {
 		super(
 			'https://polygonscan.com/',
 			'Polygon',
-			NETWORK_LIST.MATIC,
+			ChainNetwork.Matic,
 			NETWORK_IDS.MATIC,
-			'MATIC',
+			Currency.MATIC,
 			MATIC_DEPLOY,
 			maticSetts,
 		);
@@ -78,6 +80,13 @@ export const maticSetts: BadgerSett[] = [
 	},
 ];
 
-const maticTokens = maticSetts.flatMap((sett) => [sett.depositToken, sett.vaultToken]);
+export const maticRewards = [
+	{
+		address: MATIC_DEPLOY.tokens['CRV'],
+		decimals: 18,
+	},
+];
+
+const maticTokens = maticSetts.flatMap((sett) => [sett.depositToken, sett.vaultToken]).concat(maticRewards);
 
 export const maticProtocolTokens: ProtocolTokens = toRecord(maticTokens, 'address');