+++ src/mobx/stores/PricesStore.ts
@@ -10,12 +10,14 @@ import { ExchangeRates } from '../model/system-config/exchange-rates';
 import { BDiggExchangeRates } from '../model/system-config/bDigg-exchange-rates';
 import { PriceSummary } from '../model/system-config/price-summary';
 import { Network } from 'mobx/model/network/network';
+import { ChainNetwork } from '../../config/enums/chain-network.enum';
 
 export default class PricesStore {
 	private store: RootStore;
 	private priceCache: PriceSummary;
 	public exchangeRates?: ExchangeRates;
 	public bDiggExchangeRates?: BDiggExchangeRates;
+	public pricesAvailability: Record<string, boolean> = {};
 
 	constructor(store: RootStore) {
 		this.store = store;
@@ -25,6 +27,7 @@ export default class PricesStore {
 			exchangeRates: undefined,
 			bDiggExchangeRates: undefined,
 			priceCache: this.priceCache,
+			pricesAvailability: this.pricesAvailability,
 		});
 
 		observe(this.store.network, 'network', (change: IValueDidChange<Network>) => {
@@ -33,6 +36,11 @@ export default class PricesStore {
 		});
 
 		this.init();
+	}
+
+	get arePricesAvailable(): boolean {
+		const { network } = this.store.network;
+		return this.pricesAvailability[network.symbol] ?? false;
 	}
 
 	async init(): Promise<void> {
@@ -45,16 +53,22 @@ export default class PricesStore {
 	}
 
 	loadPrices = action(
-		async (network?: string): Promise<void> => {
+		async (network = ChainNetwork.Ethereum): Promise<void> => {
 			const prices = await getTokenPrices(network);
 			if (prices) {
 				Object.entries(prices).forEach((entry) => {
 					const [key, value] = entry;
 					prices[key] = new BigNumber(value);
 				});
+
 				this.priceCache = {
 					...this.priceCache,
 					...prices,
+				};
+
+				this.pricesAvailability = {
+					...this.pricesAvailability,
+					[network]: true,
 				};
 			}
 		},
@@ -82,7 +96,7 @@ export default class PricesStore {
 
 	async getBdiggExchangeRates(): Promise<BDiggExchangeRates | undefined> {
 		const url =
-			'https://api.coingecko.com/api/v3/simple/price/?ids=badger-sett-digg&vs_currencies=usd,eth,btc,cad,bnb';
+			'https://api.coingecko.com/api/v3/simple/price?ids=badger-sett-digg&vs_currencies=usd,eth,btc,cad,bnb';
 		const errorMessage = 'Failed to load exchange rates';
 		const accessor = (res: any) => res['badger-sett-digg'];
 		return fetchData(url, errorMessage, accessor);